+++ src/main/java/com/gigie/service/impl/SetmealServiceImpl.java
@@ -1,6 +1,7 @@
 package com.gigie.service.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
+import com.baomidou.mybatisplus.core.conditions.update.LambdaUpdateWrapper;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 
 import com.gigie.DTO.SetmealDto;
@@ -10,6 +11,7 @@ import com.gigie.mapper.SetmealMapper;
 import com.gigie.service.SetmealService;
 import com.gigie.service.setmealdishservice;
 
+import com.gigie.utils.CustomException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,14 +27,20 @@ public class SetmealServiceImpl extends ServiceImpl<SetmealMapper, Setmeal> impl
     private setmealdishservice setmealdishService;
     @Override @Transactional  //保存套餐和套餐关联菜品
     public void savemeal_dish(SetmealDto setmealDto) {
-        this.save(setmealDto);
+        if (!this.save(setmealDto))
+        {
+           throw new CustomException("新增套餐失败");
+        }
         List<SetmealDish> setmealDishes = setmealDto.getSetmealDishes();
         Long id=setmealDto.getId();
         for (SetmealDish setmealDish : setmealDishes) {
             setmealDish.setSetmealId(id);
         }
 
-        setmealdishService.saveBatch(setmealDishes);
+        if(!setmealdishService.saveBatch(setmealDishes))
+        {
+            throw new CustomException("新增套餐关联菜品失败");
+        }
 
     }
     /**
@@ -81,7 +89,10 @@ public class SetmealServiceImpl extends ServiceImpl<SetmealMapper, Setmeal> impl
     @Override
     @Transactional
     public void updatedish(SetmealDto setmealDto) {
-        this.updateById(setmealDto);
+        if(!this.updateById(setmealDto))
+        {
+            throw new CustomException("更新套餐失败");
+        }
         LambdaQueryWrapper<SetmealDish> queryWrapper = new LambdaQueryWrapper<>();
         queryWrapper.eq(SetmealDish ::getSetmealId,setmealDto.getId());
         setmealdishService.remove(queryWrapper);
@@ -91,7 +102,25 @@ public class SetmealServiceImpl extends ServiceImpl<SetmealMapper, Setmeal> impl
             setmealDish.setSetmealId(id);
         }
 
-        setmealdishService.saveBatch(setmealDishes);
+        if(!setmealdishService.saveBatch(setmealDishes))
+        {
+            throw new CustomException("更新套餐关联菜品失败");
+        }
 
+    }
+
+    @Autowired
+    private SetmealService SetmealService;
+    @Override
+    public void changestatus(int id, List<Long> ids) {
+
+
+        LambdaUpdateWrapper<Setmeal> queryWrapper = new LambdaUpdateWrapper<>();
+        queryWrapper.set(Setmeal ::getStatus,id);
+        queryWrapper.in(Setmeal ::getId,ids);
+        boolean update = SetmealService.update(queryWrapper);
+        if (!update) {
+            throw new CustomException("停售套餐失败");
+        }
     }
 }
